code_snippets/python/requirements.txt
restate_sdk==0.4.1
hypercorn
pydantic
requests